% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bondad.R
\name{bondad_ajuste}
\alias{bondad_ajuste}
\title{Tabla resumen de bondad de ajuste}
\usage{
bondad_ajuste(obs, sim)
}
\arguments{
\item{obs}{Vector de valores observados.}

\item{sim}{Vector de valores simulados.}
}
\value{
\code{data.frame} con NSE, R^2, RMSE y PBIAS.
}
\description{
Tabla resumen de bondad de ajuste
}
