% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ano_promedio.R
\name{mm_a_m3s}
\alias{mm_a_m3s}
\title{Conversion de lamina (mm/mes) a caudal (m3/s)}
\usage{
mm_a_m3s(mm, area, dias)
}
\arguments{
\item{mm}{Lamina mensual en mm.}

\item{area}{Area de la cuenca en km2.}

\item{dias}{Dias del mes.}
}
\value{
Caudal medio en m3/s.
}
\description{
Conversion de lamina (mm/mes) a caudal (m3/s)
}
