% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bondad.R
\name{nash_sutcliffe}
\alias{nash_sutcliffe}
\title{Eficiencia de Nash-Sutcliffe}
\usage{
nash_sutcliffe(obs, sim)
}
\arguments{
\item{obs}{Vector de valores observados.}

\item{sim}{Vector de valores simulados.}
}
\value{
Coeficiente NSE (1 = ajuste perfecto).
}
\description{
Eficiencia de Nash-Sutcliffe
}
