% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bondad.R
\name{pbias}
\alias{pbias}
\title{Sesgo porcentual (PBIAS)}
\usage{
pbias(obs, sim)
}
\arguments{
\item{obs}{Vector de valores observados.}

\item{sim}{Vector de valores simulados.}
}
\value{
PBIAS en porcentaje.
}
\description{
Sesgo porcentual (PBIAS)
}
