% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_efectiva.R
\name{pe_curva}
\alias{pe_curva}
\title{Precipitacion efectiva de una curva USBR}
\usage{
pe_curva(P, coef)
}
\arguments{
\item{P}{Vector de precipitacion mensual (mm).}

\item{coef}{Vector de 6 coeficientes polinomiales \code{c(a0, a1, a2, a3, a4, a5)}.}
}
\value{
Vector de precipitacion efectiva (mm) de la misma longitud que \code{P}.
}
\description{
Evalua el polinomio de 5 grado de una curva del metodo USBR sobre un vector de
precipitacion mensual.
}
\examples{
pe_curva(c(130.24, 90.01), coef_pe_usbr()$II)
}
