% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ano_promedio.R
\name{precip_media_mensual}
\alias{precip_media_mensual}
\title{Precipitacion media mensual a partir de una serie o un vector}
\usage{
precip_media_mensual(precip)
}
\arguments{
\item{precip}{Vector de 12 valores o \code{data.frame}/matriz ano x mes.}
}
\value{
Vector de 12 con la media mensual (enero a diciembre).
}
\description{
Precipitacion media mensual a partir de una serie o un vector
}
\keyword{internal}
