% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bondad.R
\name{r2}
\alias{r2}
\title{Coeficiente de determinacion R^2}
\usage{
r2(obs, sim)
}
\arguments{
\item{obs}{Vector de valores observados.}

\item{sim}{Vector de valores simulados.}
}
\value{
R^2 (cuadrado de la correlacion de Pearson).
}
\description{
Coeficiente de determinacion R^2
}
