% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bondad.R
\name{rmse}
\alias{rmse}
\title{Raiz del error cuadratico medio}
\usage{
rmse(obs, sim)
}
\arguments{
\item{obs}{Vector de valores observados.}

\item{sim}{Vector de valores simulados.}
}
\value{
RMSE en las unidades de los datos.
}
\description{
Raiz del error cuadratico medio
}
